# contracts/extensions/v1alpha1/provider-contracts.openapi.yaml
openapi: 3.1.0
info:
  title: Epydios AgentOps Control Plane Extension Provider Contracts
  version: v1alpha1
  summary: Versioned HTTP/JSON contracts for PolicyProvider, EvidenceProvider, and ProfileResolver.
  description: |
    This contract is the public plug-in surface used by OSS baseline providers and private
    providers (including AIMXS integrations). Breaking changes require a new version.
servers:
  - url: https://provider.internal
...
paths:
...
  /v1alpha1/profile-resolver/resolve:
    post:
      tags: [ProfileResolver]
      operationId: profileResolve
      summary: Resolve the active execution profile for a request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileResolveRequest"
      responses:
        "200":
          description: Profile selection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResolveResponse"
        "400":
          $ref: "#/components/responses/ProviderError"
        "500":
          $ref: "#/components/responses/ProviderError"
components:
  responses:
    ProviderError:
      description: Structured provider error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProviderError"
  schemas:
    ObjectMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId:
          type: string
          description: Correlation ID generated by the control plane
        timestamp:
          type: string
          format: date-time
        tenantId:
          type: string
        projectId:
          type: string
        environment:
          type: string
        actor:
          type: object
          additionalProperties: true
    ResourceRef:
      type: object
      properties:
        kind:
          type: string
        namespace:
          type: string
        name:
          type: string
        uid:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
    SubjectRef:
      type: object
      properties:
        type:
          type: string
          description: user, serviceaccount, workload, system, tenant
        id:
          type: string
        attributes:
          type: object
          additionalProperties: true
...
    ProviderError:
      type: object
      required: [errorCode, message]
      properties:
        errorCode:
          type: string
        message:
          type: string
        retryable:
          type: boolean
          default: false
        details:
          type: object
          additionalProperties: true
...
    ProfileResolveRequest:
      type: object
      required: [meta]
      properties:
        meta:
          $ref: "#/components/schemas/ObjectMeta"
        workload:
          $ref: "#/components/schemas/ResourceRef"
        subject:
          $ref: "#/components/schemas/SubjectRef"
        task:
          type: object
          properties:
            kind:
              type: string
            sensitivity:
              type: string
            latencyClass:
              type: string
            complianceTags:
              type: array
              items:
                type: string
        defaults:
          type: object
          properties:
            profileId:
              type: string
        context:
          type: object
          additionalProperties: true
    ProfileResolveResponse:
      type: object
      required: [profileId]
      properties:
        profileId:
          type: string
        profileVersion:
          type: string
        source:
          type: string
          description: e.g. static-default, policy, tenant-override, aimxs
        ttlSeconds:
          type: integer
          minimum: 0
        attributes:
          type: object
          additionalProperties: true
        requiredCapabilities:
          type: array
          items:
            type: string
        policyBundleRef:
          type: object
          properties:
            policyId:
              type: string
            policyVersion:
              type: string
            checksum:
              type: string